Actualizo constructor de codigo web de Card, y reorganizo funciones para generacion de codigo web
